--- sdk/openai/inference/src/models/response/mod.rs
@@ -1,3 +1,0 @@
-mod chat_completions;
-
-pub use chat_completions::*;
